+++ config/hashtagcmsadmin.php
@@ -5,7 +5,7 @@ return [
         'site_label' => 'CMS - Admin',
         'base_context' => 'admin',
         'base_path' => '/admin',
-        'version' => env('BE_RESOURCE_VERSION', '310120221229'),
+        'version' => env('BE_RESOURCE_VERSION', '21092023112610'),
         'theme' => 'hashtagcms::be.neo',
         'theme_assets' => 'assets/hashtagcms/be/neo',
         'app_url' => env("APP_URL"),